added new views (#142)
