+++ {{cookiecutter.project_slug}}/app.json
@@ -1,49 +1,49 @@
 {
-    "name": "{{ cookiecutter.project_name }}",
-    "env": {
-        "ALLOWED_HOSTS": {
-            "value": ".herokuapp.com"
-        },
-        "DEBUG": {
-            "value": "True"
-        },
-        "DISABLE_COLLECTSTATIC": {
-            "value": 1
-        },
-        "ENVIRONMENT": {
-            "value": "development"
-        },
-        "PROJECT_PATH": {
-            "value": "{{ cookiecutter.project_slug }}"
-        },
-        "NPM_CONFIG_PRODUCTION": {
-            "value": "false"
-        },
-        "SECRET_KEY": {
-            "generator": "secret"
-        }
+  "name": "{{ cookiecutter.project_name }}",
+  "env": {
+    "ALLOWED_HOSTS": {
+      "value": ".herokuapp.com"
     },
-    "addons": [
-        "heroku-postgresql:standard-0",
-        "papertrail:Choklad"
-    ],
-    "environments": {
-        "review": {
-            "addons": [
-                "heroku-postgresql:mini",
-                "papertrail:Choklad"
-            ]
-        }
+    "DEBUG": {
+      "value": "True"
     },
-    "buildpacks": [
-        {
-            "url": "heroku/nodejs"
-        },
-        {
-            "url": "heroku/python"
-        }
-    ],
-    "scripts": {
-        "postdeploy": "./scripts/db_setup.sh"
+    "DISABLE_COLLECTSTATIC": {
+      "value": 1
+    },
+    "ENVIRONMENT": {
+      "value": "development"
+    },
+    "PROJECT_PATH": {
+      "value": "{{ cookiecutter.project_slug }}"
+    },
+    "NPM_CONFIG_PRODUCTION": {
+      "value": "false"
+    },
+    "SECRET_KEY": {
+      "generator": "secret"
     }
+  },
+  "addons": [
+    "heroku-postgresql:standard-0",
+    "papertrail:Choklad"
+  ],
+  "environments": {
+    "review": {
+      "addons": [
+        "heroku-postgresql:mini",
+        "papertrail:Choklad"
+      ]
+    }
+  },
+  "buildpacks": [
+    {
+      "url": "heroku/nodejs"
+    },
+    {
+      "url": "heroku/python"
+    }
+  ],
+  "scripts": {
+    "postdeploy": "./scripts/db_setup.sh"
+  }
 }
