+++ {{cookiecutter.project_slug}}/server/{{cookiecutter.project_slug}}/common/favicon_urls.py
