+++ {{cookiecutter.project_slug}}/server/{{cookiecutter.project_slug}}/core/views.py
@@ -2,16 +2,7 @@ from django.conf import settings
 from django.contrib.auth import authenticate
 from django.contrib.auth.tokens import default_token_generator
 from django.db import transaction
-{% if cookiecutter.client_app == "Vue3" -%}
-from django.shortcuts import render
-from django.template.exceptions import TemplateDoesNotExist
-{% endif -%}
 from django.template.loader import render_to_string
-{% if cookiecutter.use_graphql == 'y' -%}
-from django.template.response import TemplateResponse
-from django.views.decorators.cache import never_cache
-from django.views.decorators.csrf import ensure_csrf_cookie
-{% endif -%}
 from rest_framework import generics, mixins, permissions, status, viewsets
 from rest_framework.decorators import api_view, permission_classes
 from rest_framework.exceptions import ValidationError
@@ -23,22 +14,6 @@ from .models import User
 from .permissions import CreateOnlyPermissions
 from .serializers import UserLoginSerializer, UserRegistrationSerializer, UserSerializer
 
-{% if cookiecutter.use_graphql == 'y' %}
-# Serve React frontend
-@ensure_csrf_cookie
-@never_cache
-def index(request):
-    return TemplateResponse(request, ["index.html", "core/index-placeholder.html"])
-{% elif cookiecutter.client_app.lower() == 'None' %}
-def index(request):
-    return redirect(to="/docs/swagger/")
-{% else %}
-def index(request):
-    try:
-        return render(request, "index.html")
-    except TemplateDoesNotExist:
-        return render(request, "core/index-placeholder.html")
-{% endif %}
 
 class UserLoginView(generics.GenericAPIView):
     serializer_class = UserLoginSerializer
