+++ {{cookiecutter.project_slug}}/server/{{cookiecutter.project_slug}}/urls.py
@@ -12,6 +12,6 @@ urlpatterns = [
     path("graphql", csrf_exempt(GraphQLView.as_view(graphiql=settings.DEBUG))),
     {% endif -%}
     path(r"staff/", admin.site.urls),
-    path(r"", include("{{ cookiecutter.project_slug }}.core.favicon_urls")),
-    path(r"", include("{{ cookiecutter.project_slug }}.core.urls")),
+    path(r"", include("{{ cookiecutter.project_slug }}.common.favicon_urls")),
+    path(r"", include("{{ cookiecutter.project_slug }}.common.urls")),
 ]
